refactor: make folder structure in study-path more relevant
