+++ src/features/study-path/components/stage-card/actions/actions.scss
