+++ src/features/study-path/components/stage-card/image/image.tsx
