+++ src/features/study-path/components/stage-card/image/index.ts
