+++ src/features/study-path/components/stage-card/links/links.scss
