+++ src/features/study-path/components/stage-card/logo-icon/index.ts
