+++ src/features/study-path/components/stage-card/logo-icon/logo-icon.test.tsx
