+++ src/features/study-path/components/stage-card/step/index.ts
