+++ src/features/study-path/components/stage-card/step/step.tsx
