+++ src/features/study-path/components/stage-card/topics/topics.scss
